Queries are now created: OnStartRunning
To avoid a memory leak entity queries are created during the OnStartRunning function.   The issue was OnCreate was running before Road, Connection and Auto entities were created.   This is now handled by making the Job System disabled in ER3D_Traffic while these entities are created, and then enabling the job after all that is completed.

Also a more random seed is created.